refactor(ui): consolidate overflow hover-card logic into shared hook
